+++ R/render.R
@@ -177,10 +177,10 @@ render_rtf <- function(x, display_loc = NULL, remove_unicode_ws = TRUE, use_page
   gt <- prep_obj_rtf(x)
 
   # page headers
-  gt <- apply_to_grp(
+  gt <- apply_to_gt_group(
+    gt,
     gt::tab_options,
     list(
-      data = gt,
       page.numbering = FALSE,
       page.header.use_tbl_headings = use_page_header
     )